fix: plugin bundle archive handling for upload and install
